pyproject.toml
[project]
name = "weekly-report"
version = "0.1.3"
description = "Weekly report generator using CrewAI"
requires-python = ">=3.10"

dependencies = [
  "crewai==1.7.2",
  "pydantic>=2.0",
  "mlflow>=3.8",
  "python-dotenv",
]

[project.scripts]
weekly-report = "agents_ai.weekly_report.jobs.weekly_job:main"


[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
